Code Cleanup: Move getAttributeByIds and countAttributesByGid to GermplasmService IBP-4349-BrapiUseGermplsmGUID
